apps/api/pyproject.toml
[project]
name = "stockbuddy-api"
version = "0.1.0"
description = "StockBuddy - AI-powered stock portfolio analysis API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.16",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "redis[hiredis]>=5.0.0",
    "tenacity>=8.2.0",
    "PyJWT[crypto]>=2.8.0",
    "slowapi>=0.1.9",
    "yfinance>=0.2.36",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "structlog>=24.0.0",
    "sentry-sdk[fastapi]>=1.40.0",
    "arq>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "factory-boy>=3.3.0",
    "ruff>=0.3.0",
    "aiosqlite>=0.20.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
